+++ src/main/java/com/java/project/turek/TurekService.java
@@ -18,7 +18,7 @@ public class TurekService {
         return turekList.stream().map(TurekDTO::new).toList();
     }
 
-    public TurekDTO getTurekById(Long id) {
+    public TurekDTO getTurekById(int id) {
         return new TurekDTO(turekRepository.findById(id).orElseThrow(()->new RuntimeException("Turek not found")));
     }
 
@@ -26,7 +26,7 @@ public class TurekService {
         return turekRepository.save(turek);
     }
 
-    public Turek updateTurek(Turek turek, Long id) {
+    public Turek updateTurek(Turek turek, int id) {
     Turek existingTurek = turekRepository.findById(id).orElseThrow(() -> new RuntimeException("Turek not found"));
     existingTurek.setImie(turek.getImie());
     existingTurek.setNazwisko(turek.getNazwisko());
@@ -37,11 +37,11 @@ public class TurekService {
     return turekRepository.save(existingTurek);
 }
 
-    public void deleteTurek(Long id) {
+    public void deleteTurek(int id) {
     turekRepository.deleteById(id);
 }
 
-    public Turek deactivateTurek(Long id) {
+    public Turek deactivateTurek(int id) {
     Turek turek = turekRepository.findById(id).orElseThrow(() -> new RuntimeException("Turek not found"));
     turek.setStatus(0);
     return turekRepository.save(turek);
